+++ resources/views/home.blade.php
@@ -116,7 +116,7 @@
                             <td>Otto</td>
                             <td>@mdo</td>
                             <td><a href="/events/edit"><i class="fa-regular fa-pen-to-square"></i></a></td>
-                            <td><a href="/events/delete"><i class="fa-sharp fa-solid fa-trash" style="color: #db1414;"></i></a></td>
+                            <td><a href="/events/delete" class="confirm-delete-link" data-toggle="modal" data-target="#confirmDeleteModal" data-item-id="1"><i class="fa-sharp fa-solid fa-trash" style="color: #db1414;"></i></a></td>
                         </tr>
                         <tr>
                             <th scope="row">3</th>
@@ -126,7 +126,7 @@
                             <td>Otto</td>
                             <td>@mdo</td>
                             <td><a href="/events/edit"><i class="fa-regular fa-pen-to-square"></i></a></td>
-                            <td><a href="/events/delete"><i class="fa-sharp fa-solid fa-trash" style="color: #db1414;"></i></a></td>
+                            <td><a href="/events/delete" class="confirm-delete-link" data-toggle="modal" data-target="#confirmDeleteModal" data-item-id="1"><i class="fa-sharp fa-solid fa-trash" style="color: #db1414;"></i></a></td>
                         </tr>
                         <tr>
                             <th scope="row">4</th>
@@ -136,7 +136,7 @@
                             <td>Otto</td>
                             <td>@mdo</td>
                             <td><a href="/events/edit"><i class="fa-regular fa-pen-to-square"></i></a></td>
-                            <td><a href="/events/delete"><i class="fa-sharp fa-solid fa-trash" style="color: #db1414;"></i></a></td>
+                            <td><a href="/events/delete" class="confirm-delete-link" data-toggle="modal" data-target="#confirmDeleteModal" data-item-id="1"><i class="fa-sharp fa-solid fa-trash" style="color: #db1414;"></i></a></td>
                         </tr>
                     @endslot
                     <x-confirm-delete-modal>
